Multi level job change set
When there are multi layer jobs, the leaf job is not able to find the
change set. As the change set is not available though it calculates the
score, it is not able to credit the points to the player. Now we
recursively go to the top most upstream job and add it to the list of
accountable builds causing the leaf job to find the change set.